Refactor to_unit into amount module
